[FLINK-19722][table-planner] separate table sources, table sinks, source functions, sink functions and lookup functions from TestValuesTableFactory and TestValuesTableRuntimeFunctions to single classes
